omit unstated & typescriptize
